bdproj/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
